move just a bunch of things around for no reason
