+++ client/blocks/factions.js
@@ -3,7 +3,7 @@ import {withTracker} from 'meteor/react-meteor-data';
 import formJson from '@quarterto/form-json';
 import {Cards} from '../../shared/collections';
 import Ornamented from '../components/ornamented';
-import Icon from '../components/icon';
+import Icon from '../components/presentation/icon';
 import styled from 'styled-components';
 import {withCampaign} from '../components/campaign';
 import {compose, withHandlers, withProps} from 'recompose';
