--- client/components/icon.js
@@ -1,5 +1,0 @@
-import React from 'react';
-
-const Icon = ({icon, className}) => <i className={`ra ra-${icon} ${className}`} />;
-
-export default Icon;
